feat(auth-module): update the auth module and remove teh role and permissions entities
