+++ src/app.module.ts
@@ -2,8 +2,6 @@ import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { SharedModule } from './shared/shared.module';
 import { AuthModule } from './modules/auth/auth.module';
-import { RoleModule } from './modules/role/role.module';
-import { PermissionModule } from './modules/permission/permission.module';
 import { CompanyModule } from './modules/company/company.module';
 import { UserModule } from './modules/user/user.module';
 import { IsUnique } from './shared/util/validator/is-unique-validator.util';
@@ -19,8 +17,6 @@ import { EntityModule } from './modules/entity/entity.module';
     }),
     SharedModule,
     AuthModule,
-    RoleModule,
-    PermissionModule,
     CompanyModule,
     UserModule,
     EntityModule,
