--- src/modules/permission/permission.controller.ts
@@ -1,7 +1,0 @@
-import { Controller } from '@nestjs/common';
-import { PermissionService } from './permission.service';
-
-@Controller('permission')
-export class PermissionController {
-  constructor(private readonly permissionService: PermissionService) {}
-}
