--- src/modules/permission/permission.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import { PermissionService } from './permission.service';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { Permission } from './entities/permission.entity';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([Permission])],
-  providers: [PermissionService],
-})
-export class PermissionModule {}
