--- src/modules/role/dto/create-role.dto.ts
@@ -1,1 +1,0 @@
-export class CreateRoleDto {}
