--- src/modules/role/dto/update-role.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateRoleDto } from './create-role.dto';
-
-export class UpdateRoleDto extends PartialType(CreateRoleDto) {}
